Extract ManagedArtifact into separate file
